+++ src/main/java/com/cardano/explorer/model/response/BlockResponse.java
@@ -1,6 +1,6 @@
 package com.cardano.explorer.model.response;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
@@ -23,9 +23,9 @@ public class BlockResponse {
 
   private Integer epochSlotNo;
 
-  private BigDecimal totalFees;
+  private BigInteger totalFees;
 
-  private BigDecimal totalOutput;
+  private BigInteger totalOutput;
 
   private String slotLeader;
