+++ src/main/java/com/cardano/explorer/model/response/EpochResponse.java
@@ -1,7 +1,7 @@
 package com.cardano.explorer.model.response;
 
 import com.cardano.explorer.common.enumeration.EpochStatus;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
@@ -16,7 +16,7 @@ public class EpochResponse {
 
   private Integer blkCount;
 
-  private BigDecimal outSum;
+  private BigInteger outSum;
 
   private Integer txCount;
