+++ src/main/java/com/cardano/explorer/model/response/pool/projection/EpochStakeProjection.java
@@ -1,10 +1,10 @@
 package com.cardano.explorer.model.response.pool.projection;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public interface EpochStakeProjection {
 
   Integer getEpochNo();
 
-  BigDecimal getTotalStake();
+  BigInteger getTotalStake();
 }
